% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming_history.R
\name{convert_to_df}
\alias{convert_to_df}
\title{Finds Top Artist you listen to
Finds Top Artist you listen to}
\usage{
convert_to_df(filepath)
}
\arguments{
\item{filepath}{JSON file}
}
\value{
top playlists
}
\description{
Finds Top Artist you listen to
Finds Top Artist you listen to
}
