% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamviz.R
\name{month_detect}
\alias{month_detect}
\title{Manipulates users data to be used for visualization}
\usage{
month_detect(df)
}
\arguments{
\item{df}{The data frame converted from JSON file}
}
\value{
A dataframe
}
\description{
Manipulates users data to be used for visualization
}
\examples{
dataframe <- month_detect(userdf)

}
