% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamviz.R
\name{plot_top_artists}
\alias{plot_top_artists}
\title{Visualize your most listened-to artists by minutes listened}
\usage{
plot_top_artists(filepath = NULL, dat = NULL, n = 5)
}
\arguments{
\item{filepath}{JSON file}

\item{dat}{a tidy data frame}

\item{n}{the number of artists to return}
}
\value{
graph of minutes played for entire year
}
\description{
Visualize your most listened-to artists by minutes listened
}
