% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming_history.R
\name{top_artists}
\alias{top_artists}
\title{Finds Top Artist you listen to}
\usage{
top_artists(filepath, n = 5)
}
\arguments{
\item{filepath}{JSON file}
}
\value{
A dataframe with top artists
}
\description{
Finds Top Artist you listen to
}
