% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamviz.R
\name{top_artists_time_graph}
\alias{top_artists_time_graph}
\title{Find your most listened-to artists by minutes listened}
\usage{
top_artists_time_graph(filepath = NULL, dat = NULL, n = n)
}
\arguments{
\item{filepath}{JSON file}

\item{dat}{a tidy data frame}

\item{n}{the number of artists to return}
}
\value{
graph of minutes played for entire year
}
\description{
Find your most listened-to artists by minutes listened
}
