% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming_history.R
\name{top_songs}
\alias{top_songs}
\title{Find your most listened-to songs}
\usage{
top_songs(filepath, n = 5)
}
\arguments{
\item{filepath}{JSON file}
}
\value{
A dataframe with top songs
}
\description{
Find your most listened-to songs
}
