% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamviz.R
\name{year_wrap}
\alias{year_wrap}
\title{Creates a visualization for minutes played for the entire year}
\usage{
year_wrap(df)
}
\arguments{
\item{df}{Manipulated dataframe}
}
\value{
graph of minutes played for entire year
}
\description{
Creates a visualization for minutes played for the entire year
}
\examples{
df <- month_detect(userdf)
year_wrap(df)

}
