% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming_history.R
\name{convert_to_df}
\alias{convert_to_df}
\title{Converts JSON file to a tidy dataframe}
\usage{
convert_to_df(filepath)
}
\arguments{
\item{filepath}{The path to the json file}
}
\value{
A tidy dataframe
}
\description{
Converts JSON file to a tidy dataframe
}
