% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filterStream.R
\name{createOAuthToken}
\alias{createOAuthToken}
\title{Create OAuth token without handshake.}
\usage{
createOAuthToken(consumerKey, consumerSecret, accessToken, accessTokenSecret)
}
\description{
Create OAuth token without handshake.
}
\details{
This function generates a OAuth token using the consumer key, consumer secret,
access token and access token secret available in the "Keys and Access Token"
tab of the "Application Management" website on Twitter's developers website.
}

