% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamR-package.R
\docType{data}
\name{example_tweets}
\alias{example_tweets}
\title{Ten sample tweets published by @twitterapi}
\source{
\url{http://www.twitter.com/twitterapi}
}
\usage{
data(example_tweets)
}
\description{
A vector of string characters that contains ten sample tweets
in plain text.
}
\keyword{datasets}
