% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamviz.R
\name{month_wrap}
\alias{month_wrap}
\title{Visualize your distribution of minutes listened for the month}
\usage{
month_wrap(filepath = NULL, dat = NULL, dmonth)
}
\arguments{
\item{dmonth}{string of month user would like to visualize, "01 - 12"}

\item{df}{Manipulated dataframe}
}
\value{
graph of minutes played for selected month
}
\description{
Visualize your distribution of minutes listened for the month
}
