% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamviz.R
\name{month_wrap}
\alias{month_wrap}
\title{Creates a visualization for minutes played based on users input}
\usage{
month_wrap(filepath = NULL, dat = NULL, dmonth)
}
\arguments{
\item{dmonth}{string of month user would like to visualize, "01 - 12"}

\item{df}{Manipulated dataframe}
}
\value{
graph of minutes played for wanted month
}
\description{
Creates a visualization for minutes played based on users input
}
\examples{

month_wrap(df, "05")

}
