\docType{package}
\name{streamR-package}
\alias{streamR}
\alias{streamR-package}
\title{Access to Twitter Streaming APIs via R}
\description{
  This package provides a series of functions that allow R
  users to access Twitter's filter, sample, and user
  streams, and to parse the output into data frames.
}
\author{
  Pablo Barbera \email{pablo.barbera@nyu.edu}
}
\seealso{
  \code{\link{filterStream}}, \code{\link{sampleStream}},
  \code{\link{userStream}}, \code{\link{readTweets}},
  \code{\link{parseTweets}}
}

