% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaming_history.R
\name{top_artists}
\alias{top_artists}
\title{Find your most listened-to artists by number of streams}
\usage{
top_artists(filepath = NULL, dat = NULL, n = 5)
}
\arguments{
\item{filepath}{JSON file}

\item{dat}{a tidy data frame}

\item{n}{the number of artists to return}
}
\value{
A dataframe with top artists
}
\description{
Find your most listened-to artists by number of streams
}
