% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_artists.R
\name{top_artists}
\alias{top_artists}
\title{Find your most listened-to artists by number of streams or minutes listened}
\usage{
top_artists(filepath = NULL, dat = NULL, method = "streams", n = 5)
}
\arguments{
\item{filepath}{path to a JSON file}

\item{dat}{a tidy data frame}

\item{method}{how to find top artists ("streams" or "minutes")}

\item{n}{the number of artists to return}
}
\value{
A dataframe with top artists
}
\description{
Find your most listened-to artists by number of streams or minutes listened
}
