% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamviz.R
\name{year_wrap}
\alias{year_wrap}
\title{Visualize your distribution of minutes listened for the year}
\usage{
year_wrap(filepath = NULL, dat = NULL)
}
\arguments{
\item{filepath}{path to a JSON file}

\item{dat}{a tidy data frame}
}
\value{
graph of minutes played in the past year
}
\description{
Visualize your distribution of minutes listened for the year
}
\examples{
df <- month_detect(userdf)
year_wrap(df)

}
